Refactoring, added documentation
