+++ hawc/apps/animal/api.py
@@ -7,8 +7,13 @@ from rest_framework.decorators import action
 from rest_framework.exceptions import NotAcceptable, PermissionDenied
 from rest_framework.response import Response
 
-from ..assessment.api import AssessmentLevelPermissions, AssessmentViewset, DoseUnitsViewset
-from ..common.api import CleanupFieldsBaseViewSet, user_can_edit_object
+from ..assessment.api import (
+    AssessmentLevelPermissions,
+    AssessmentViewset,
+    CleanupFieldsBaseViewSet,
+    DoseUnitsViewset,
+    user_can_edit_object,
+)
 from ..common.constants import AssessmentViewsetPermissions
 from ..common.helper import FlatExport, re_digits
 from ..common.renderers import PandasRenderers
