+++ hawc/apps/assessment/api/__init__.py
@@ -1,3 +1,4 @@
 from .filters import *  # noqa: F401,F403
+from .helper import *  # noqa: F401,F403
 from .permissions import *  # noqa: F401,F403
 from .viewsets import *  # noqa: F401,F403
