+++ tests/hawc/apps/common/test_api.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.test import APIClient
 
-from hawc.apps.common.api import user_can_edit_object
+from hawc.apps.assessment.api import user_can_edit_object
 from hawc.apps.common.diagnostics import worker_healthcheck
 from hawc.apps.myuser.models import HAWCUser
 from hawc.apps.study.models import Study
